Updates for VS2022.

+++ TMG/g_utils.c
@@ -382,12 +382,12 @@ char *G_CopyString (char *in)
 }
 
 
-void G_InitEdict (edict_t *e)
+void G_InitEdict(edict_t* e)
 {
 	e->inuse = true;
 	e->classname = "noclass";
 	e->gravity = 1.0;
-	e->s.number = (int) (e - g_edicts);
+	e->s.number = (e - g_edicts);
 
 	// Clear what the free-edict list may have set.
 	e->chain = NULL;
